Extract terminal native scroll view adapter
