+++ openspec/changes/reduce-macos-architecture-debt-warnings/tasks.md
@@ -22,11 +22,11 @@
 
 ## 4. Terminal Surface And Host Debt
 
-- [ ] 4.1 Identify the next `TerminalSurfaceController.swift` adapter or input/surface boundary that can move without changing behavior.
-- [ ] 4.2 Extract the selected terminal surface boundary and run focused terminal surface validation.
+- [x] 4.1 Identify the next `TerminalSurfaceController.swift` adapter or input/surface boundary that can move without changing behavior.
+- [x] 4.2 Extract the selected terminal surface boundary and run focused terminal surface validation.
 - [ ] 4.3 Identify the next `TerminalHostView.swift` collaborator boundary for runtime attachment, overlay presentation, input routing, window observation, metadata publishing, or surface coordination.
 - [ ] 4.4 Extract the selected terminal host boundary and run focused terminal host/runtime validation.
-- [ ] 4.5 Update `clients/apple/ARCHITECTURE.md` and script expectations after each reduced warning.
+- [x] 4.5 Update `clients/apple/ARCHITECTURE.md` and script expectations after each reduced warning.
 - [ ] 4.6 Commit and open stacked PRs for each reduced terminal warning.
 
 ## 5. Console Content View Debt
